+++ jcm/physics/speedy/speedy_physics.py
@@ -64,7 +64,7 @@ class SpeedyPhysics(Physics):
         from jcm.physics.clouds.speedy_condensation import get_large_scale_condensation_tendencies
         from jcm.physics.radiation.speedy_shortwave import get_shortwave_rad_fluxes, get_clouds
         from jcm.physics.radiation.speedy_longwave import get_downward_longwave_rad_fluxes, get_upward_longwave_rad_fluxes
-        from jcm.physics.speedy.surface_flux import get_surface_fluxes
+        from jcm.physics.surface.speedy_surface_flux import get_surface_fluxes
         from jcm.physics.speedy.vertical_diffusion import get_vertical_diffusion_tend
         from jcm.physics.speedy.forcing import set_forcing
         # from jcm.physics.speedy.orographic_correction import get_orographic_correction_tendencies
